+++ rust/src/trade/mod.rs
@@ -10,15 +10,16 @@ mod types;
 pub use context::{EstimateMaxPurchaseQuantityResponse, SubmitOrderResponse, TradeContext};
 pub use push_types::{PushEvent, PushOrderChanged, TopicType};
 pub use requests::{
-    EstimateMaxPurchaseQuantityOptions, GetCashFlowOptions, GetFundPositionsOptions,
-    GetHistoryExecutionsOptions, GetHistoryOrdersOptions, GetStockPositionsOptions,
-    GetTodayExecutionsOptions, GetTodayOrdersOptions, ReplaceOrderOptions, SubmitOrderOptions,
+    EstimateMaxPurchaseQuantityOptions, GetAllExecutionsOptions, GetCashFlowOptions,
+    GetFundPositionsOptions, GetHistoryExecutionsOptions, GetHistoryOrdersOptions,
+    GetStockPositionsOptions, GetTodayExecutionsOptions, GetTodayOrdersOptions,
+    ReplaceOrderOptions, SubmitOrderOptions,
 };
 pub use types::{
-    AccountBalance, BalanceType, CashFlow, CashFlowDirection, CashInfo, ChargeCategoryCode,
-    CommissionFreeStatus, DeductionStatus, Execution, FrozenTransactionFee, FundPosition,
-    FundPositionChannel, FundPositionsResponse, MarginRatio, Order, OrderChargeDetail,
-    OrderChargeFee, OrderChargeItem, OrderDetail, OrderHistoryDetail, OrderSide, OrderStatus,
-    OrderTag, OrderType, OutsideRTH, StockPosition, StockPositionChannel, StockPositionsResponse,
-    TimeInForceType, TriggerPriceType, TriggerStatus,
+    AccountBalance, AllExecutionsResponse, BalanceType, CashFlow, CashFlowDirection, CashInfo,
+    ChargeCategoryCode, CommissionFreeStatus, DeductionStatus, Execution, FrozenTransactionFee,
+    FundPosition, FundPositionChannel, FundPositionsResponse, MarginRatio, Order,
+    OrderChargeDetail, OrderChargeFee, OrderChargeItem, OrderDetail, OrderHistoryDetail, OrderSide,
+    OrderStatus, OrderTag, OrderType, OutsideRTH, StockPosition, StockPositionChannel,
+    StockPositionsResponse, TimeInForceType, TriggerPriceType, TriggerStatus,
 };